Issue #3986: Splitting inputs by Test for AnonInnerLength
